The Pit Scouting Page Works
There is a separate tab, for those who are logged in, to look up a team
number and then edit their scouting page. When they click submit, the
page is redirected to the teams page and all the information from pit
scouting can be viewed on the bottom of the page.